% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rc-package.R
\docType{package}
\name{fastrc-package}
\alias{fastrc}
\alias{fastrc-package}
\title{fastrc: Fast Reverse Complement of DNA and RNA Sequences}
\description{
Fast reverse complement of DNA and RNA sequences using a C++ lookup table for O(1) per-base complement mapping. Supports full IUPAC ambiguity codes, DNA and RNA modes, case preservation, and NA handling. Much faster than other packages for computing reverse complements of many short sequences such as primers, probes, and, k-mers.
}
\author{
\strong{Maintainer}: Steven G. Rozen \email{steverozen@pm.me}

}
\keyword{internal}
