% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rc-package.R
\docType{package}
\name{fastrc-package}
\alias{fastrc}
\alias{fastrc-package}
\title{fastrc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
